+++ selenium/src/main/java/org/musetest/selenium/providers/FirefoxMarionetteDriverProvider.java
@@ -2,6 +2,7 @@ package org.musetest.selenium.providers;
 
 import org.musetest.core.*;
 import org.musetest.core.events.*;
+import org.musetest.core.util.*;
 import org.musetest.selenium.*;
 import org.openqa.selenium.*;
 import org.openqa.selenium.firefox.*;
@@ -13,45 +14,37 @@ import java.io.*;
  * @author Christopher L Merrill (see LICENSE.txt for license details)
  */
 @MuseTypeId("firefox-marionette-provider")
-public class FirefoxMarionetteDriverProvider implements WebDriverProvider
+public class FirefoxMarionetteDriverProvider extends BaseLocalDriverProvider
     {
     @Override
     public WebDriver getDriver(SeleniumBrowserCapabilities capabilities, MuseExecutionContext context)
         {
-        synchronized (FirefoxMarionetteDriverProvider.class)
+        if (getOs() != null && !(OperatingSystem.get().equals(getOs())))
+            return null;   // this provider is not for the current OS
+
+        if (!capabilities.getCapabilities().get(SeleniumBrowserCapabilities.BROWSER_NAME).equals(BrowserType.FIREFOX))
+            return null;
+
+        File path = getDriverLocation(context);
+        if (path == null)
             {
-            if (capabilities.getCapabilities().get(SeleniumBrowserCapabilities.BROWSER_NAME).equals(BrowserType.FIREFOX))
-                {
-                if (_path_to_exe == null)
-                    {
-                    context.raiseEvent(new MessageEvent("FirefoxMarionetteDriverProvider would try to satisfy request for Firefox browser, but it was not provided with a path-to-exe"));
-                    return null;
-                    }
-
-                if (!(new File(_path_to_exe).exists()))
-                    {
-                    context.raiseEvent(new MessageEvent("FirefoxMarionetteDriverProvider would try to satisfy request for Firefox browser, but the provided path-to-exe does not exist"));
-                    return null;
-                    }
-
-                System.setProperty("webdriver.gecko.driver", _path_to_exe);
-                DesiredCapabilities selenium_capabilities = capabilities.toDesiredCapabilities();
-                selenium_capabilities.setCapability("marionette", true);
-                return new MarionetteDriver(selenium_capabilities);
-                }
+            context.raiseEvent(new MessageEvent("FirefoxMarionetteDriverProvider would try to satisfy request for Firefox browser, but it was not configured with a path to the driver"));
             return null;
             }
-        }
 
-    public String getPathToExe()
-        {
-        return _path_to_exe;
-        }
+        if (!(path.exists()))
+            {
+            context.raiseEvent(new MessageEvent("FirefoxMarionetteDriverProvider would try to satisfy request for Firefox browser, but the configured path does not exist: " + path.getAbsolutePath()));
+            return null;
+            }
 
-    @SuppressWarnings("unused")  // required for Json de/serialization
-    public void setPathToExe(String path_to_exe)
-        {
-        _path_to_exe = path_to_exe;
+        synchronized (FirefoxMarionetteDriverProvider.class)
+            {
+            System.setProperty("webdriver.gecko.driver", path.getAbsolutePath());
+            DesiredCapabilities selenium_capabilities = capabilities.toDesiredCapabilities();
+            selenium_capabilities.setCapability("marionette", true);
+            return new MarionetteDriver(selenium_capabilities);
+            }
         }
 
     @Override
@@ -59,8 +52,6 @@ public class FirefoxMarionetteDriverProvider implements WebDriverProvider
         {
         return "MarionetteDriver";
         }
-
-    private String _path_to_exe;
     }