+++ selenium/src/main/java/org/musetest/selenium/providers/SafariDriverProvider.java
@@ -1,6 +1,7 @@
 package org.musetest.selenium.providers;
 
 import org.musetest.core.*;
+import org.musetest.core.util.*;
 import org.musetest.selenium.*;
 import org.openqa.selenium.*;
 import org.openqa.selenium.remote.*;
@@ -10,16 +11,20 @@ import org.openqa.selenium.safari.*;
  * @author Christopher L Merrill (see LICENSE.txt for license details)
  */
 @MuseTypeId("safaridriver-provider")
-public class SafariDriverProvider implements WebDriverProvider
+public class SafariDriverProvider extends BaseLocalDriverProvider
     {
     @Override
     public WebDriver getDriver(SeleniumBrowserCapabilities capabilities, MuseExecutionContext context)
         {
+        if (getOs() != null && !(OperatingSystem.get().equals(getOs())))
+            return null;   // this provider is not for the current OS
+
+        if (!capabilities.getCapabilities().get(SeleniumBrowserCapabilities.BROWSER_NAME).equals(BrowserType.SAFARI))
+            return null;
+
         synchronized (SafariDriverProvider.class)
             {
-            if (capabilities.getCapabilities().get(SeleniumBrowserCapabilities.BROWSER_NAME).equals(BrowserType.SAFARI))
-                return new SafariDriver(capabilities.toDesiredCapabilities());
-            return null;
+            return new SafariDriver(capabilities.toDesiredCapabilities());
             }
         }