+++ change-log.md
@@ -3,6 +3,8 @@
 ## 0.6
 - Minor changes in support of the first IDE release
 - switch to LogBack (from slf4j-simple) for logging diagnostic. Log to a file instead of console ($HOME/.muse/muse.log) 
+- enhance: Support relative and absolute paths for local driver providers
+- enhance: Support OS-specific configurations for local driver providers
 
 ## 0.5
 - refactor: resource lookup from project return ResourceTokens, instead of the resources. The tokens may then be used to fetch the resource.